Added seperate thread option, changes for python3
Calling Console.ConsoleApp() from the interpreter (or a script) will start the console in a new thread and launch the window while things continue running. The problem is that during the execution of code in the new console all output from the original program will also be re-directed to the console.

Added other changes to make compatible with python3, but it isn't working yet.